+++ packages/core/src/components/dropdown-menu/dropdown-menu.tsx
@@ -64,7 +64,7 @@ export class DropdownMenu {
     options.setAttribute("aria-labelledby", this.button.id);
 
     for (const ul of Array.from(this.host.getElementsByTagName("ul"))) {
-      ul.setAttribute("role", "none");
+      ul.setAttribute("role", "group");
       for (const li of Array.from(ul.getElementsByTagName("li"))) {
         li.setAttribute("role", "none");
       }
@@ -75,10 +75,9 @@ export class DropdownMenu {
     for (const li of Array.from(this.host.getElementsByTagName("li"))) {
       for (const tab of tabbable(li)) {
         tab.setAttribute("role", this.checkable ? "menuitemradio" : "menuitem");
-        if (this.checkable && li.classList.contains("dso-checked")) {
-          tab.setAttribute("aria-checked", "true");
-        } else {
-          tab.removeAttribute("aria-checked");
+
+        if (this.checkable) {
+          tab.setAttribute("aria-checked", li.classList.contains("dso-checked").toString());
         }
       }
     }